+++ _meta.lua
@@ -2,5 +2,5 @@ local _ = require("gettext")
 return {
     fullname = _("komix"),
     description = _([[Browse, search, download and sync your Komga library from KOReader.]]),
-    version = "1.0.6",
+    version = "1.0.7",
 }